+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 #########
 
+3.0.0
+*****
+
+- ``Collection::map()`` now passes both indexes and values to the mapper callback
+- added ``Map::build()`` , ``Map::apply()``, ``Map::intersectKeys()``, ``Map::uintersectKeys()``,
+  ``Map::diffKeys()``, ``Map::udiffKeys()``
+
+
 2.0.0
 *****